+++ engine/src/flutter/impeller/renderer/backend/gles/capabilities_gles.h
@@ -157,7 +157,7 @@ class CapabilitiesGLES final
   ISize GetMaximumRenderPassAttachmentSize() const override;
 
   // |Capabilities|
-  float GetMaxSamplerAnisotropy() const override;
+  uint32_t GetMaxSamplerAnisotropy() const override;
 
   // |Capabilities|
   size_t GetMinimumUniformAlignment() const override;
@@ -179,7 +179,7 @@ class CapabilitiesGLES final
   bool supports_texture_compression_etc2_ = false;
   bool supports_texture_compression_astc_ = false;
   bool supports_texture_compression_astc_hdr_ = false;
-  float max_sampler_anisotropy_ = 1.0f;
+  uint32_t max_sampler_anisotropy_ = 1;
   PixelFormat default_glyph_atlas_format_ = PixelFormat::kUnknown;
 };
