+++ engine/src/flutter/lib/gpu/context.cc
@@ -154,7 +154,7 @@ extern bool InternalFlutterGpu_Context_GetSupportsManuallyMippedTextures(
       ->SupportsManuallyMippedTextures();
 }
 
-extern double InternalFlutterGpu_Context_GetMaxSamplerAnisotropy(
+extern int InternalFlutterGpu_Context_GetMaxSamplerAnisotropy(
     flutter::gpu::Context* wrapper) {
   return wrapper->GetContext().GetCapabilities()->GetMaxSamplerAnisotropy();
 }
